divide file
